2022_python/pyproject.toml
[tool.poetry]
name = "2022-python"
version = "0.1.0"
description = ""
authors = ["Tyler Vick <vick.tyler@gmail.com>"]
readme = "README.md"
packages = [{include = "2022_python"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.5"
scipy = "^1.9.3"
python-dotenv = "^0.21.0"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
